requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1;python_version<"3.9"
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0.6
django-allauth==0.51.0
django-heroku==0.3.1
django-simpleui==2022.12.18
Flask==2.1.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
future==0.18.2
googlemaps==4.6.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
python-slugify==7.0.0
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
Werkzeug==2.2.1
whitenoise==6.2.0
wincertstore==0.2
yarl==1.7.2
zipp==3.8.1
stripe==5.0.0
